Add a custom field-type `DBFocusPoint`. Move all focus point calculations into this class. Remove unneeded InterventionBackend.
