+++ _config/config.yml
@@ -17,6 +17,4 @@ SilverStripe\Admin\LeftAndMain:
     - 'jonom/focuspoint: client/css/FocusPointField.css'
   extra_requirements_javascript:
     - 'jonom/focuspoint: client/javascript/FocusPointField.js'
-# SilverStripe\Core\Injector\Injector:
-#   SilverStripe\Assets\Image_Backend:
-#     class: JonoM\FocusPoint\InterventionBackend
+
