+++ tests/FocusPointTest.yml
@@ -8,29 +8,29 @@ SilverStripe\Assets\Image:
     FileFilename: folder/pngTestImage.png
     FileHash: 9c2e4494af6f4b01a590d82dbaa44f6be8b63f05
     Parent: =>SilverStripe\Assets\Folder.folder1
-    FocusX: -.25
-    FocusY: .25
+    FocusPointFocusX: -.25
+    FocusPointFocusY: .25
   pngRightTop:
     Title: PNG w right/top focus point
     Name: pngTestImage.png
     FileFilename: folder/pngTestImage.png
     FileHash: 9c2e4494af6f4b01a590d82dbaa44f6be8b63f05
     Parent: =>SilverStripe\Assets\Folder.folder1
-    FocusX: .25
-    FocusY: .25
+    FocusPointFocusX: .25
+    FocusPointFocusY: .25
   pngRightBottom:
     Title: PNG w right/bottom focus point
     Name: pngTestImage.png
     FileFilename: folder/pngTestImage.png
     FileHash: 9c2e4494af6f4b01a590d82dbaa44f6be8b63f05
     Parent: =>SilverStripe\Assets\Folder.folder1
-    FocusX: .25
-    FocusY: -.25
+    FocusPointFocusX: .25
+    FocusPointFocusY: -.25
   pngLeftBottom:
     Title: PNG w left/bottom focus point
     Name: pngTestImage.png
     FileFilename: folder/pngTestImage.png
     FileHash: 9c2e4494af6f4b01a590d82dbaa44f6be8b63f05
     Parent: =>SilverStripe\Assets\Folder.folder1
-    FocusX: -.25
-    FocusY: -.25
+    FocusPointFocusX: -.25
+    FocusPointFocusY: -.25
